fold diverginging implementations
